// src/enums/StorageMechanism.ts
// src/enums/StorageMechanism.ts

export enum StorageMechanism {
  /**
   * Local storage mechanism.
   */
  LOCAL = 'local',

  /**
   * Session storage mechanism.
   */
  SESSION = 'session',

  /**
   * Cookie storage mechanism.
   */
  COOKIE = 'cookie',

  /**
   * Memory storage mechanism.
   */
  MEMORY = 'memory',
}